Added include_rules preprocessor functionality.
This allows for including of rules from other files. This works
recursively. An error was identified that needs to be addressed, where
if there is no semicolon terminating the final statement (as in, the
last valid token in the source file needs to be a semicolon), then the
system will crash with a vector out of range error. The culprit is the
current line 649, and additional checks need to be put in to stop that
error.