Started searching for and fixing parser problems.
Implementation of the parsing language pre-parser is not perfect, and
needs to be rigorously tested to ensure that the pre-parser is parsing
the parsing language source code correctly. Strongly consider
refactoring the way the if statements are structured to simplify them,
as they are currently a bit obtuse and unwieldly.