Added new lexer to accept string literals.
The new lexing algorithm, denoted by tokenizeFile2, is identical to
tokenizeFile except that it also accepts arbitrarily complex string
literals (a token comprised of a starting and ending " character, with
the enclosed contents being anything of any length, including whitespace
and other characters that would normally count as delimiters in the
lexing algorithm). Additionally, an erroneous design decision was fixed:
The last commit saw the user being able to define a lexing algorithm of
their own, which is great, except in the code it was being used to lex
the parsing langauge source file, which is incorrect. It has now been
fixed to only be used to lex the source file that is to be parsed by the
parsing algorithm.