Began work in moving to use of string literals.
The lexer now provides the parser parser with arbitrarily complex string
literals where needed as single tokens. Because string literals are now
single tokens and not three tokens ( " token " ), checks need to be
changed to look for entire string tokens instead of single ". Work is
underway in this respect, but by no means completed.