Implemented a terrible similar-string algorithm.
This new algorithm is in use at the step where tokens that are not
literal strings, nor operators, nor keywords, nor the names of rules are
detected. It is used to provide a list of rule-names the programmer may
have meant. I suppose it could be changed to also detect whether it was
similar to a keyword.