Initiated large changes to design of parsing language.
The change in design is focused around essentially completely removing the complex symbol
table system, in favor of a simple create-scope/delete-scope system for simple variable
name usage checks, as well as adding simple named lists that can be referenced against
and added to. The parser, with this branch commit, now either fully or partially
recognizes this new syntax, while having removed most or all of the obsolete syntax from
the previous design.