Modified the text transform function to be semi-useful.
Moreover, there is so much spaghetti in both the parser parser and the design of this
language that serious consideration should be given to rewriting the parser parser,
and solidifying the syntax and grammar of the parsing language before reimplementing
the parser parser or implementing the parser interpreter. Of particular note are the
contextual uses of !, the effect of * and whether + (from formal PEGs) needs to be
implemented or if special uses of * cover it, and whether & (again from formal PEGs)
needs to be implemented, and how (it probably does need to be implemented). Basically
just need to sit down and carve everything into stone, then semi-start-over.